+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/magic/MageController.java
@@ -593,4 +593,5 @@ public interface MageController {
 
     @Nullable
     MagicAttribute getAttribute(String attributeKey);
+    boolean isUndoable(@Nonnull Material material);
 }