+++ Magic/CHANGELOG.md
@@ -9,6 +9,7 @@
  - Prevent putting wands into a grindstone
  - Fix wands disappearing under certain circumstances when dropped.
    However, they will drop now under those same circumstances. It's the best I could do!
+ - Add "undoable" material list, for tweaking the undo system (e.g. excluding fire from undo)
 
 # 7.9.7